requirements.txt
pytest==8.3.2
playwright==1.46.0
allure-pytest==2.13.5